refactor(agents): separate prepared snapshot construction
Move unchanged snapshot construction beside its full-catalog owner and
currentness assertions beside their errors. Preserve current catalog-attempt
reporting, initialization and per-run auth-store creation.

Related: #140674
